Moved AsmAnnotator and AsmCompletionContributor label handling to a shared cache
