+++ src/main/java/com/x366/asm/AsmCompletionContributor.java
@@ -3,17 +3,14 @@ package com.x366.asm;
 import com.intellij.codeInsight.completion.*;
 import com.intellij.codeInsight.lookup.LookupElementBuilder;
 import com.intellij.openapi.editor.Document;
+import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.patterns.PlatformPatterns;
 import com.intellij.util.ProcessingContext;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class AsmCompletionContributor extends CompletionContributor {
-
-    private static final Pattern LABEL_PATTERN = Pattern.compile("^\\s*([a-zA-Z_]\\w*):", Pattern.MULTILINE);
 
     private static final List<String> KEYWORDS = AsmLexer.KEYWORD_SET.stream().sorted().toList();
     private static final List<String> REGISTERS = AsmLexer.REGISTER_SET.stream().sorted().toList();
@@ -50,9 +47,12 @@ public class AsmCompletionContributor extends CompletionContributor {
                         for(String reg : REGISTERS) {
                             result.addElement(LookupElementBuilder.create(reg).withTypeText("register"));
                         }
-                        Matcher matcher = LABEL_PATTERN.matcher(doc.getText());
-                        while(matcher.find()) {
-                            result.addElement(LookupElementBuilder.create(matcher.group(1)).withTypeText("label").withTailText(":"));
+
+                        VirtualFile vFile = parameters.getOriginalFile().getVirtualFile();
+                        if(vFile != null) {
+                            for(String label : AsmLabelCache.getLabels(vFile.getPath(), doc)) {
+                                result.addElement(LookupElementBuilder.create(label).withTypeText("label").withTailText(":"));
+                            }
                         }
                     }
                 }
